baum_one/requirements.txt
flask~=3.0.0
fastapi~=0.109.0
pydantic~=2.5.3
SQLAlchemy~=2.0.25
uvicorn~=0.25.0
python-dotenv~=1.0.0
alembic~=1.13.1
asyncpg~=0.29.0
python-multipart~=0.0.6
gunicorn~=21.2.0
fastapi-filter~=1.1.0
pytest~=8.0.0
flake8~=7.0.0
isort~=5.13.2
aiokafka~=0.10.0
bitstring~=4.1.4
kafka-python~=2.0.2